Cleanup and prepare auth middleware for rbauth
