+++ server/controllers/alert.controller.js
@@ -1,5 +1,5 @@
 const { createController } = require("awilix-express");
-const { ensureAuthenticated } = require("../middleware/auth");
+const { authenticate } = require("../middleware/authenticate");
 const { AppConstants } = require("../server.constants");
 const { validateInput } = require("../handlers/validators");
 const { idRules } = require("./validation/generic.validation");
@@ -8,21 +8,17 @@ const {
   createAlertRules,
   updateAlertRules
 } = require("./validation/alert-controller.validation");
-const Logger = require("../handlers/logger.js");
 
 class AlertController {
-  #serverVersion;
-  #serverPageTitle;
   #settingsStore;
   #alertService;
   #scriptService;
 
-  logger = new Logger("Server-API");
+  #logger;
 
-  constructor({ settingsStore, serverVersion, alertService, scriptService, serverPageTitle }) {
+  constructor({ settingsStore, alertService, scriptService, loggerFactory }) {
     this.#settingsStore = settingsStore;
-    this.#serverVersion = serverVersion;
-    this.#serverPageTitle = serverPageTitle;
+    this.#logger = loggerFactory("Server-API");
     this.#alertService = alertService;
     this.#scriptService = scriptService;
   }
@@ -68,7 +64,7 @@ class AlertController {
 // prettier-ignore
 module.exports = createController(AlertController)
     .prefix(AppConstants.apiRoute + "/alert")
-    .before([ensureAuthenticated])
+    .before([authenticate])
     .get("/", "list")
     .post("/", "create")
     .put("/:id", "update")
