+++ server/controllers/settings/storage.controller.js
@@ -1,5 +1,5 @@
 const { createController } = require("awilix-express");
-const { ensureAuthenticated } = require("../../middleware/auth");
+const { authenticate } = require("../../middleware/authenticate");
 const Logger = require("../../handlers/logger.js");
 const multer = require("multer");
 const { AppConstants } = require("../../server.constants");
@@ -27,5 +27,5 @@ class StorageController {
 // prettier-ignore
 module.exports = createController(StorageController)
   .prefix(AppConstants.apiRoute + "/settings")
-  .before([ensureAuthenticated])
+  .before([authenticate])
   .post("/backgroundUpload", "uploadBackgroundImage", {before: [upload.single("myFile")]});
